+++ image_processing.py
@@ -1,4 +1,29 @@
 import cv2
+import numpy as np
+
+def get_only_instance_image(input_file, mask, width, height, output_file=None, show=False):
+	'''
+	파일 이름을 입력받아서 실제로 masking 된 부분의 사진만 output 해 준다.
+	input_file	: string, 파일 이름
+	masks		: 2차원 list, width / height 크기, True 면 그 자리에 객체가 있는것, 아니면 없음.
+	width		: int, 너비
+	height		: int, 높이
+	output_file	: string, output_file 이름
+	'''
+	if output_file == None:
+		output_file = input_file.split(".")[0] + "_masked" + input_file.split(".")[1]
+	original = cv2.imread(input_file)
+	masked_image = np.zeros([height, width ,3], dtype=np.uint8)
+	mask_num = 0
+
+	for h in range(0, height):
+		for w in range(0, width):
+				for c in range(0, 3):
+					masked_image[h][w][c] = (original[h][w][c] if mask[h][w] else 0)
+					if mask[h][w]:
+						mask_num += 1
+
+	return masked_image, mask_num
 
 def get_contours(frame, start=190, diff=30):
 	'''
@@ -18,30 +43,28 @@ def get_contours(frame, start=190, diff=30):
 	# Using Contours, search Contours.
 	return cv2.findContours(edges_high_thresh, cv2.RETR_LIST,cv2.CHAIN_APPROX_NONE)
 
-def contours_to_coord(contours):
+def get_average_color(image, total_class, tc_num):
 	'''
-	입력받은 외곽선들을 좌표계로 바꾸어서 return 해 준다.
-	contours : 2D List. 각 원소들은 외곽선의 sequence 이다.
+	image 의 주어진 tc_num개의 total_class에서의 평균 색을 구한다.
 	'''
-	coords = []
-	for cinstances in contours:
-		temp = []
-		for c in cinstances:
-				now = (c[0][0], c[0][1])
-				if now not in temp:
-					temp.append(now)
-		coords.append(temp)
-	return coords
+	aver_r = 0
+	aver_g = 0
+	aver_b = 0
 
-def delete_line_threshold(contours, line_n=40):
-	'''
-	입력받은 외곽선 중, line_n 갯수를 넘지 않는 작은 line들은 지워준다.
-	'''
-	contour_len = list(map(len, contours))
+	for coord in total_class:
+		rgb = image[coord[1]][coord[0]]
+		aver_r += rgb[0]
+		aver_g += rgb[1]
+		aver_b += rgb[2]
+	
+	return [int(aver_r/tc_num), int(aver_g/tc_num), int(aver_b/tc_num)]
 
-	ret_contours = []
-
-	for i in range(0, len(contours)):
-		if contour_len[i] > line_n:
-			ret_contours.append(contours[i])
-	return ret_contours
+def get_class_color(image, class_total, class_count, color_function=get_average_color):
+	'''
+	class total의 list에 있는 각각의 평균적인 색상을 image에서 구한다.
+	color_function 으로 색을 구하는 방식도 지정 가능하다.
+	'''
+	class_color = []
+	for i in range(len(class_total)):
+		class_color.append(color_function(image, class_total[i], class_count[i]))
+	return class_color
